Split datepicker component into core and implementation
